fix: 修复声音OSD上有些数值显示不对的问题。
程序内部将double类型乘100转换为int类型出现问题，此为数据类型转换问题。

Log: 修复声音OSD上有些数值显示不对的问题。
Bug: https://pms.uniontech.com/bug-view-154561.html
Influence: 无影响。

+++ dde-osd/src/common.cpp
@@ -94,7 +94,7 @@ void DrawHelper::DrawBackground(QPainter *painter, const QStyleOptionViewItem &o
     painter->fillPath(path, QColor("#2ca7f8"));
 }
 
-void DrawHelper::DrawPercentValue(QPainter *painter, const QStyleOptionViewItem &option, int value, const QColor &color)
+void DrawHelper::DrawPercentValue(QPainter *painter, const QStyleOptionViewItem &option, double value, const QColor &color)
 {
     const int rightMargin = 16;
     const int percentFlagWidth = 9; // %号宽度
